refactor(openapi): Refine Schema decorators
