+++ openapi/src/decorators/path.ts
@@ -1,5 +1,5 @@
-import {getOpenApiMeta, Param, ParamLocation, ParamOptions, RequestBody, ResponseDescriptor} from "../meta";
-import {Schema} from "../helpers";
+import { getOpenApiMeta } from "../meta";
+import { ParamLocation, ParamOptions, ResponseDescriptor, SchemaDef } from "../types";
 
 function basicDecoratorFactory(key: string, value: any): MethodDecorator {
   return (target: any, method: string, descriptor: any) => {
@@ -15,9 +15,6 @@ export function Summary(summary: string): MethodDecorator {
 }
 export function Description(description: string): MethodDecorator {
   return basicDecoratorFactory('description', description);
-}
-export function Parameters(params: Param[]): MethodDecorator {
-  return basicDecoratorFactory('parameters', params);
 }
 export function Param(name: string, location: ParamLocation, options: ParamOptions = {}): MethodDecorator {
   return (target: any, method: string, descriptor: any) => {
@@ -39,10 +36,7 @@ export function Deprecated(urls?: string[]): MethodDecorator {
   if (urls) return basicDecoratorFactory('deprecated', urls);
   return basicDecoratorFactory('deprecated', true);
 }
-export function RequestBody(bodyDefinition: RequestBody): MethodDecorator {
-  return basicDecoratorFactory('requestBody', bodyDefinition);
-}
-export function BodyContent(mediaType: string, schema: Schema): MethodDecorator {
+export function BodyContent(mediaType: string, schema: SchemaDef): MethodDecorator {
   return (target: any, method: string, descriptor: any) => {
     const meta = getOpenApiMeta(target);
     const methodMeta = meta[method] = meta[method] || {};
@@ -55,8 +49,8 @@ export function Responses(def: { [key: string]: ResponseDescriptor }): MethodDec
   return basicDecoratorFactory('responses', def);
 }
 export function OpenApiResponse(status: string | number, description: string): MethodDecorator;
-export function OpenApiResponse(status: string | number, produces: string, schema: Schema): MethodDecorator;
-export function OpenApiResponse(status: string | number, descriptionOrProduces: string, schema?: Schema): MethodDecorator {
+export function OpenApiResponse(status: string | number, produces: string, schema: SchemaDef): MethodDecorator;
+export function OpenApiResponse(status: string | number, descriptionOrProduces: string, schema?: SchemaDef): MethodDecorator {
   return (target: any, method: string, descriptor: any) => {
     const meta = getOpenApiMeta(target);
     const methodMeta = meta[method] = meta[method] || {};
