+++ openapi/src/index.ts
@@ -1,2 +1,4 @@
-export * from './helpers';
 export * from './decorators';
+export * from './helpers';
+export * from './types';
+
